requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.3
bootstrap4==0.1.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.6
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.4
django-crispy-forms==2.3
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.67.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyDrive==1.3.1
pyparsing==3.2.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
